requirements.txt
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
streamlit>=1.30.0
google-generativeai>=0.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
plotly>=5.5.0
openpyxl>=3.0.9
psycopg2-binary>=2.9.3
mysql-connector-python>=8.0.28
pymssql>=2.2.5
sqlite3-api>=0.1.0
